Break up re_smart_channel in modules
