+++ crates/re_smart_channel/src/lib.rs
@@ -1,13 +1,19 @@
 //! A channel that keeps track of latency and queue length.
 
 use std::sync::{
-    atomic::{AtomicBool, AtomicU64, Ordering::Relaxed},
+    atomic::{AtomicBool, AtomicU64},
     Arc,
 };
 
 use web_time::Instant;
 
 pub use crossbeam::channel::{RecvError, RecvTimeoutError, SendError, TryRecvError};
+
+mod receiver;
+mod sender;
+
+pub use receiver::Receiver;
+pub use sender::Sender;
 
 // --- Source ---
 
@@ -122,7 +128,7 @@ impl std::fmt::Display for SmartMessageSource {
 
 /// Stats for a channel, possibly shared between chained channels.
 #[derive(Default)]
-struct SharedStats {
+pub(crate) struct SharedStats {
     /// Latest known latency from sending a message to receiving it, it nanoseconds.
     latency_ns: AtomicU64,
 }
@@ -138,7 +144,7 @@ pub fn smart_channel<T: Send>(
 /// Create a new channel using the same stats as some other.
 ///
 /// This is a very leaky abstraction, and it would be nice to refactor some day
-fn smart_channel_with_stats<T: Send>(
+pub(crate) fn smart_channel_with_stats<T: Send>(
     sender_source: SmartMessageSource,
     source: SmartChannelSource,
     stats: Arc<SharedStats>,
@@ -160,95 +166,6 @@ fn smart_channel_with_stats<T: Send>(
 }
 
 // ---
-
-#[derive(Clone)]
-pub struct Sender<T: Send> {
-    tx: crossbeam::channel::Sender<SmartMessage<T>>,
-    source: Arc<SmartMessageSource>,
-    stats: Arc<SharedStats>,
-}
-
-impl<T: Send> Sender<T> {
-    /// Clones the sender with an updated source.
-    pub fn clone_as(&self, source: SmartMessageSource) -> Self {
-        Self {
-            tx: self.tx.clone(),
-            source: Arc::new(source),
-            stats: Arc::clone(&self.stats),
-        }
-    }
-
-    pub fn send(&self, msg: T) -> Result<(), SendError<T>> {
-        self.send_at(
-            Instant::now(),
-            Arc::clone(&self.source),
-            SmartMessagePayload::Msg(msg),
-        )
-    }
-
-    /// Forwards a message as-is.
-    pub fn send_at(
-        &self,
-        time: Instant,
-        source: Arc<SmartMessageSource>,
-        payload: SmartMessagePayload<T>,
-    ) -> Result<(), SendError<T>> {
-        // NOTE: We should never be sending a message with an unknown source.
-        debug_assert!(!matches!(*source, SmartMessageSource::Unknown));
-
-        self.tx
-            .send(SmartMessage {
-                time,
-                source,
-                payload,
-            })
-            .map_err(|SendError(msg)| SendError(msg.into_data().unwrap()))
-    }
-
-    /// Used to indicate that a sender has left.
-    ///
-    /// This sends a message down the channel allowing the receiving end to know whether one of the
-    /// sender has left, and if so why (if applicable).
-    ///
-    /// Using a [`Sender`] after calling `quit` is undefined behavior: the receiving end is free
-    /// to silently drop those messages (or worse).
-    pub fn quit(
-        &self,
-        err: Option<Box<dyn std::error::Error + Send>>,
-    ) -> Result<(), SendError<SmartMessage<T>>> {
-        // NOTE: We should never be sending a message with an unknown source.
-        debug_assert!(!matches!(*self.source, SmartMessageSource::Unknown));
-
-        self.tx.send(SmartMessage {
-            time: Instant::now(),
-            source: Arc::clone(&self.source),
-            payload: SmartMessagePayload::Quit(err),
-        })
-    }
-
-    /// Is the channel currently empty of messages?
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.tx.is_empty()
-    }
-
-    /// Number of messages in the channel right now.
-    #[inline]
-    pub fn len(&self) -> usize {
-        self.tx.len()
-    }
-
-    /// Latest known latency from sending a message to receiving it, it nanoseconds.
-    pub fn latency_ns(&self) -> u64 {
-        self.stats.latency_ns.load(Relaxed)
-    }
-
-    /// Latest known latency from sending a message to receiving it,
-    /// in seconds
-    pub fn latency_sec(&self) -> f32 {
-        self.latency_ns() as f32 / 1e9
-    }
-}
 
 /// The payload of a [`SmartMessage`].
 ///
@@ -295,129 +212,6 @@ impl<T: Send> SmartMessage<T> {
             Msg(msg) => Some(msg),
             Quit(_) => None,
         }
-    }
-}
-
-pub struct Receiver<T: Send> {
-    rx: crossbeam::channel::Receiver<SmartMessage<T>>,
-    stats: Arc<SharedStats>,
-    source: SmartChannelSource,
-    connected: AtomicBool,
-}
-
-impl<T: Send> Receiver<T> {
-    /// Are we still connected?
-    ///
-    /// Once false, we will never be connected again: the source has run dry.
-    ///
-    /// This is only updated once one of the receive methods fails.
-    pub fn is_connected(&self) -> bool {
-        self.connected.load(Relaxed)
-    }
-
-    pub fn recv(&self) -> Result<SmartMessage<T>, RecvError> {
-        let msg = match self.rx.recv() {
-            Ok(x) => x,
-            Err(RecvError) => {
-                self.connected.store(false, Relaxed);
-                return Err(RecvError);
-            }
-        };
-
-        let latency_ns = msg.time.elapsed().as_nanos() as u64;
-        self.stats.latency_ns.store(latency_ns, Relaxed);
-
-        Ok(msg)
-    }
-
-    pub fn try_recv(&self) -> Result<SmartMessage<T>, TryRecvError> {
-        let msg = match self.rx.try_recv() {
-            Ok(x) => x,
-            Err(err) => {
-                if err == TryRecvError::Disconnected {
-                    self.connected.store(false, Relaxed);
-                }
-                return Err(err);
-            }
-        };
-
-        let latency_ns = msg.time.elapsed().as_nanos() as u64;
-        self.stats.latency_ns.store(latency_ns, Relaxed);
-
-        Ok(msg)
-    }
-
-    pub fn recv_timeout(
-        &self,
-        timeout: std::time::Duration,
-    ) -> Result<SmartMessage<T>, RecvTimeoutError> {
-        let msg = match self.rx.recv_timeout(timeout) {
-            Ok(x) => x,
-            Err(err) => {
-                if err == RecvTimeoutError::Disconnected {
-                    self.connected.store(false, Relaxed);
-                }
-                return Err(err);
-            }
-        };
-
-        let latency_ns = msg.time.elapsed().as_nanos() as u64;
-        self.stats.latency_ns.store(latency_ns, Relaxed);
-
-        Ok(msg)
-    }
-
-    /// Receives without registering the latency.
-    ///
-    /// This is for use with [`Sender::send_at`] when chaining to another channel
-    /// created with [`Self::chained_channel`].
-    pub fn recv_with_send_time(&self) -> Result<SmartMessage<T>, RecvError> {
-        self.rx.recv()
-    }
-
-    /// Where is the data coming from?
-    #[inline]
-    pub fn source(&self) -> &SmartChannelSource {
-        &self.source
-    }
-
-    /// Is the channel currently empty of messages?
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.rx.is_empty()
-    }
-
-    /// Number of messages in the channel right now.
-    #[inline]
-    pub fn len(&self) -> usize {
-        self.rx.len()
-    }
-
-    /// Latest known latency from sending a message to receiving it, it nanoseconds.
-    pub fn latency_ns(&self) -> u64 {
-        self.stats.latency_ns.load(Relaxed)
-    }
-
-    /// Latest known latency from sending a message to receiving it,
-    /// in seconds
-    pub fn latency_sec(&self) -> f32 {
-        self.latency_ns() as f32 / 1e9
-    }
-
-    /// Create a new channel that use the same stats as this one.
-    ///
-    /// This means both channels will see the same latency numbers.
-    ///
-    /// Care must be taken to use [`Self::recv_with_send_time`] and [`Sender::send_at`].
-    /// This is a very leaky abstraction, and it would be nice with a refactor.
-    pub fn chained_channel(&self) -> (Sender<T>, Receiver<T>) {
-        smart_channel_with_stats(
-            // NOTE: We cannot know yet, and it doesn't matter as the new sender will only be used
-            // to forward existing messages.
-            SmartMessageSource::Unknown,
-            self.source.clone(),
-            self.stats.clone(),
-        )
     }
 }
